+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Release Notes
 
+## Version v0.0.4 - 2020-06-19
+
+Features:
+
+- Added swagger-ui deploy
+- Added swagger-editor deploy
+
+Improvements:
+
+- Improvements function deploy
+
 ## Version v0.0.3 - 2020-06-10
 
 Features: